requirements.txt
fastapi==0.135.1
fitz==0.0.1.dev2
numpy==2.4.3
pandas==3.0.1
Pillow==12.1.1
scikit_learn==1.8.0
sentence_transformers==5.2.3
starlette==0.52.1
uvicorn==0.41.0
